requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
pydantic-settings
rich
asgi-correlation-id
asgi_correlation_id
python-json-logger
logtail-python
python-jose
passlib[bcrypt]
python-multipart
pywhatkit